REDDEV-539 config update
The previous definition which produced the configuration option wasn't correct.

{
  "key": "chart-definitions",
  "type": "json-array",
  "value": "[]"
}

should have been,

{
  "name": "Chart Definitions",
  "key": "chart-definitions",
  "type": "json-array"
}

however that would allow someone to enter a chart definition manually. I'm

I refactored to include a descriptive piece of text for when someone tries
to configure the module. There's two places to configure the module just like
everything else. Once in Control Center, which is where you'll find the
drop-down shown in the issue screen shot labeled, "Module configuration permissions in projects:".

The second place is in the Project Setup > External Modules page. That configure button
shows these project settings. I'm not sure that we have any for the module like we did
for the plugin where we had to configure the config project pid.

I'm still looking into how the permissions work, but won't have a chance to talk with Liz
until tomorrow. What's on dev right now is broken, so this fixes it with some placeholder
text and still enables that drop-down for configuring permissions.