REDDEV-539 removed check for module access
This reverts to the previous permission model.

A user can only edit charts if they are a super user or have design rights.

For a user to view charts they must have design rights, or have the Vizr
module checked on the User Rights page. For the later, the module must
be configured to 'Require module-specific user privilege'.