REDDEV-637 Allow non-standard external module locations
This fix accomodates instances where the external modules location is
not the default location. When Vizr was a plugin it was dependent on the
`redcap_connect.php` file to be in a known location, but the external
module API automatically provides this file so it can be removed as a
dependency in the PHP files.